Update administrator page

+++ BachHoaOnline/Controllers/AccountController.cs
@@ -13,16 +13,29 @@ namespace BachHoaOnline.Controllers
 
         public IActionResult Index()
         {
+            if (HttpContext.Session.Get<string>("user") != "admin")
+            {
+                return RedirectToAction("Error","Administrator");
+            }
             return View(db.Khachhang.ToList());
         }
 
         public IActionResult Create()
         {
+            if (HttpContext.Session.Get<string>("user") != "admin")
+            {
+                return RedirectToAction("Error", "Administrator");
+            }
             return View();
         }
 
         public IActionResult DoCreate(Khachhang kh)
         {
+            if (HttpContext.Session.Get<string>("user") != "admin")
+            {
+                return RedirectToAction("Error", "Administrator");
+            }
+
             db.Khachhang.Add(kh);
             ViewBag.status = "Tạo mới thành công.";
             return View("Create");
@@ -30,11 +43,21 @@ namespace BachHoaOnline.Controllers
 
         public IActionResult Update(int id)
         {
+            if (HttpContext.Session.Get<string>("user") != "admin")
+            {
+                return RedirectToAction("Error", "Administrator");
+            }
+
             return View(db.Khachhang.Where(x => x.Makh == id).SingleOrDefault());
         }
 
         public IActionResult DoUpdate(Khachhang kh)
         {
+            if (HttpContext.Session.Get<string>("user") != "admin")
+            {
+                return RedirectToAction("Error", "Administrator");
+            }
+
             Khachhang des = db.Khachhang.Where(x => x.Makh == kh.Makh).SingleOrDefault();
             des.Hoten = kh.Hoten;
             des.Matkhau = kh.Matkhau;
@@ -49,12 +72,22 @@ namespace BachHoaOnline.Controllers
         }
 
         public IActionResult Delete(int id)
-        {            
+        {
+            if (HttpContext.Session.Get<string>("user") != "admin")
+            {
+                return RedirectToAction("Error", "Administrator");
+            }
+
             return View(db.Khachhang.Where(x => x.Makh == id).SingleOrDefault());
         }
 
         public IActionResult DoDelete(Khachhang kh)
         {
+            if (HttpContext.Session.Get<string>("user") != "admin")
+            {
+                return RedirectToAction("Error", "Administrator");
+            }
+
             db.Khachhang.Remove(db.Khachhang.Where(x => x.Makh == kh.Makh).SingleOrDefault());
             db.SaveChanges();
             ViewBag.status = "Đã xóa thành công.";
